Set up to use a robot to deploy
We're going to have Travis automatically build our site whenever a new commit
to master comes through.  It will push to the gh-pages branch.

Right now the Travis tool does not support deploy keys, and instead use personal
access tokens.  This seems pretty insecure, but the workaround is to create a user
that ONLY exists for this task on github.  That is now done.

Have Travis actually build things